LayoutTests/http/tests/websocket/tests/hybi/contentextensions/upgrade-worker-expected.txt

CONSOLE MESSAGE: Content blocker promoted URL from ws://127.0.0.1/websocket/tests/hybi/simple to wss
CONSOLE MESSAGE: WebSocket network error: The operation couldn’t be completed. Connection refused
CONSOLE MESSAGE: line 21: PASS worker onclose url wss://127.0.0.1/websocket/tests/hybi/simple
